Move functions to another file.
